Change stucture of files, correct files
