--- config.py
@@ -1,30 +1,0 @@
-# Конфигурация и константы проекта
-
-# Отношение состояний сделки
-STATUSES_LEADS = {
-    "Нет оплаты": 66431750,
-    "Оплачена полностью": 142,
-    "Оплачена частично": 142,
-    "Оплачена с переплатой": 142,
-    "Отменена": 143,
-}
-
-# ID воронки "Воронка" в проекте amoCRM
-ID_VOR = 8117898
-
-
-data = [
-    {
-        "airtable_id": "M13752",
-        "amount_paid": 7000,
-        "course_code": "ISB_7.7.07",
-        "date_received": "2021-10-01",
-        "price": 7000,
-        "student_email": "uuu@gmail.com",
-        "student_name": "Name1",
-        "student_phone": "+770000000000",
-        "student_surname": "Surname1",
-        "submission_id": "P18814",
-        "status": "Оплачена полностью",
-    },
-]
