+++ flask_amocrm_project/config.py
@@ -1,0 +1,13 @@
+# Конфигурация и константы проекта
+
+# Отношение состояний сделки
+STATUSES_LEADS = {
+    "Нет оплаты": 66431750,
+    "Оплачена полностью": 142,
+    "Оплачена частично": 142,
+    "Оплачена с переплатой": 142,
+    "Отменена": 143,
+}
+
+# ID воронки "Воронка" в проекте amoCRM
+ID_VOR = 8117898
