+++ flask_amocrm_project/requirements.txt
@@ -3,4 +3,4 @@ requests==2.31.0
 Flask==3.0.3
 Jinja2==3.1.4
 MarkupSafe==2.1.5
-Werkzeug==3.0.3
+Werkzeug==3.1.3
