+++ requirements-dev.txt
@@ -1,7 +1,1 @@
-certifi==2024.8.30
-charset-normalizer==3.4.0
-idna==3.10
 pre-commit==3.6.0
-python-dotenv==1.0.0
-requests==2.31.0
-urllib3==2.2.3
